+++ tests/Parsers/JsonapiParserTest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Swis\JsonApi\Client\Tests\JsonApi;
+namespace Swis\JsonApi\Client\Tests\Parsers;
 
 use Art4\JsonApiClient\Utils\Manager;
 use Swis\JsonApi\Client\Jsonapi;
-use Swis\JsonApi\Client\JsonApi\JsonapiParser;
-use Swis\JsonApi\Client\JsonApi\MetaParser;
 use Swis\JsonApi\Client\Meta;
+use Swis\JsonApi\Client\Parsers\JsonapiParser;
+use Swis\JsonApi\Client\Parsers\MetaParser;
 use Swis\JsonApi\Client\Tests\AbstractTest;
 
 class JsonapiParserTest extends AbstractTest